# internal/api/docs/swagger.yaml
basePath: /api
info:
  contact: {}
  description: VoHive 主服务控制面 API — 自动生成 OpenAPI 文档 POC
  title: VoHive Main API
  version: "1.0"
paths:
...
  /settings/notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: GetNotificationSettings
      tags:
      - settings
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: UpdateNotificationSettings
      tags:
      - settings
  /settings/notifications/bark/test:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: TestBarkNotification
      tags:
      - settings
  /settings/notifications/email/test:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: TestEmailNotification
      tags:
      - settings
  /settings/notifications/webhook/test:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: TestWebhookNotification
      tags:
      - settings
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"